app/requirements.txt
fastapi
uvicorn
google-cloud-dialogflow-cx
google-cloud-speech
google-cloud-texttospeech
python-multipart
gunicorn